Rename sensor.position → .track_point, so position can be absolute
We want to be able to calculate absolute positions of sensors so that we
can click on them. We'll do this by hooking up to the piece_positioned
signal and setting sensor.position relative to that.

This renaming therefore moves the sensor's TrackPoint attribute out of
the way, and makes our naming more consistent (i.e. a .position
attribute should always be a Position instance)